Take out archive tests into their own file
